fixed order of doc vs. params for marginalia
The order of arguments to defn was wrong. The first parameter
can be either a string (documentation) or a vector. The second
or third argument respectively is the first form evaluated. If
the first form is a string this does not affect the function,
but on the other hand doesn't set the correct doc meta information
that is required by marginalia.